+++ frontend-service/my-app/src/ButtonGrid.css
@@ -1,0 +1,49 @@
+body {
+    background-color: #ffe1bc;
+}
+  
+code {
+    font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+      monospace;
+}
+
+#Title{
+    text-align: center;
+}
+
+.Grid-Container {
+    display: flex;
+    flex-direction:column;
+    justify-content: space-evenly;
+    align-items: center;
+    height: 10rem; 
+}
+
+.row {
+    display: flex;
+    flex-direction:row;
+    justify-content: space-evenly;
+    width: 10rem;
+}
+
+button
+{
+    
+    display: inline-block;
+    outline: none;
+    cursor: pointer;
+    font-size: 16px;
+    line-height: 20px;
+    font-weight: 600;
+    border-radius: 8px;
+    padding: 14px 24px;
+    border: none;
+    transition: box-shadow 0.2s ease 0s, -ms-transform 0.1s ease 0s, -webkit-transform 0.1s ease 0s, transform 0.1s ease 0s;
+    background: linear-gradient(to right, rgb(230, 30, 77) 0%, rgb(227, 28, 95) 50%, rgb(215, 4, 102) 100%);
+    color: #fff;
+
+}
+
+
+
+
