+++ frontend-service/my-app/src/index.tsx
@@ -3,14 +3,13 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import Component from './Component';
+import Component from './ButtonGrid';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
 root.render(
   <React.StrictMode>
-    {/* <App /> */}
     <Component/>
   </React.StrictMode>
 );
